LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/size-query-with-var-expected.txt


PASS Matching --match-unknown
FAIL Matching --match-unknown-fallback assert_equals: expected "yes" but got "no"
FAIL Matching --match-unregistered assert_equals: expected "yes" but got "no"
PASS Matching --match-unregistered-keyword
PASS Matching --match-unregistered-number
FAIL Matching --match-registered assert_equals: expected "yes" but got "no"
PASS Matching --match-registered-keyword
PASS Matching --match-registered-number
